moved a bunch of stuff out into modules instead of inheritable classes. will do a writeup of reasoning
