+++ lib/adjective/concerns/imbibable.rb
@@ -20,20 +20,11 @@ module Adjective
       end
     end
 
+    # Level down functionality should work the same way as simply setting a level.
+
     def normalize_experience
       @experience = 0 if @experience < 0
     end        
-
-    # def level_down
-    #   if experience < @active_exp_set[@level]
-    #     @level -=1
-    #     return true
-    #   else
-    #     return false
-    #   end
-    # end 
-    #  0  1   2   3   4   5   6   7   8    9    10
-    # [0,200,300,400,500,600,700,800,900,1000, 1200]
 
     def max_level
       # Essentially dropping the 0 index so we can access directly with an integer, not int - 1
@@ -68,42 +59,7 @@ module Adjective
     def experience_to_next_level
       return nil if max_level?
       return @active_exp_set[@level+1] - @experience
-    end
-
-    # def grant_experience(exp_to_grant, opts = {})
-    #   # Only takes positive integers - should avoid bugs and underflow this way.
-    #   if exp_to_grant < 0
-    #     raise RuntimeError, "[#{Time.now}]: Provided value in #grant_experience (#{exp_to_grant}) is not a positive integer."
-    #   else
-    #     @experience += exp_to_grant
-    #     normalize_experience
- 
-    #     if !opts[:suppress_levels]
-    #       until !can_level_up?
-    #         level_up
-    #       end
-    #     end
-
-    #     return { :exp_granted => exp_to_grant, :total_exp => @experience }
-    #   end   
-    # end
-
-    # def subtract_experience(exp_to_subtract, opts = {})
-    #   # Only takes positive integers - should avoid bugs and underflow this way.
-    #   if exp_to_subtract < 0
-    #     raise RuntimeError, "#{Time.now}]: Provided value in #subtract_experience (#{exp_to_subtract}) is not a positive integer."
-    #   else
-    #     @experience -= exp_to_subtract
-    #     normalize_experience
-
-    #     if !opts[:suppress_levels]
-    #       until !level_down
-    #         level_down
-    #       end
-    #     end
-
-    #   end 
-    # end   
+    end  
 
   end
