+++ spec/inventory_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Adjective::Inventory do
+RSpec.describe Adjective::Storable do
 
   before(:example) do
     # Clear globals as we need to track specific items by their instance_id to test them appropriately.
@@ -15,13 +15,13 @@ RSpec.describe Adjective::Inventory do
     @quiver = SurrogateItem.new({id: 5,  name: "Quiver"}) #7
     @empty_quiver = SurrogateItem.new({id: 5,  name: "Quiver", ammunition: 0}) #8
 
-    @inventory = Adjective::Inventory.new([@item, @item, @item, @item])
-    @diverse_inventory = Adjective::Inventory.new([@item, @stick, @wool, @item, @stick, @wool])
-    @extended_inventory = Adjective::Inventory.new([@full_mana_potion, @full_health_potion, @partial_health_potion, @quiver, @empty_quiver ])
-    @mixed_attribute_inventory = Adjective::Inventory.new([@item, @item, @quiver, @stick, @full_health_potion])
+    @inventory = SurrogateInventory.new("Backpack", 1, [@item, @item, @item, @item])
+    @diverse_inventory = SurrogateInventory.new("Backpack", 1, [@item, @stick, @wool, @item, @stick, @wool])
+    @extended_inventory = SurrogateInventory.new("Backpack", 1, [@full_mana_potion, @full_health_potion, @partial_health_potion, @quiver, @empty_quiver ])
+    @mixed_attribute_inventory = SurrogateInventory.new("Backpack", 1, [@item, @item, @quiver, @stick, @full_health_potion])
 
     @child_inventory = SurrogateInventory.new("Backpack", 1, [@full_health_potion, @partial_health_potion, @stick, @wool, @quiver])
-    @limited_inventory = Adjective::Inventory.new([@item, @item, @quiver, @stick, @full_health_potion], {max_size: 8})
+    @limited_inventory = SurrogateInventory.new("Backpack", 1, [@item, @item, @quiver, @stick, @full_health_potion], {max_size: 8})
   end
 
   context "when using a child model" do 
@@ -37,20 +37,20 @@ RSpec.describe Adjective::Inventory do
   context "when initialized" do 
 
     it "will accept no items as an argument" do 
-      expect(Adjective::Inventory.new().items.length).to eq(0)
+      expect(SurrogateInventory.new().items.length).to eq(0)
     end
 
     it "will accept an empty items array as an argument" do 
-      expect(Adjective::Inventory.new([]).items.length).to eq(0)
+      expect(SurrogateInventory.new([]).items.length).to eq(0)
     end    
 
     it "will accept an array argument of items" do 
-      inventory = Adjective::Inventory.new([@item, @item])
+      inventory = SurrogateInventory.new("Pack", 2, [@item, @item])
       expect(inventory.items.length).to be(2)
     end
 
     it "will set max_size to :unlimited by default" do
-      default_inventory = Adjective::Inventory.new
+      default_inventory = SurrogateInventory.new
       expect(default_inventory.max_size).to eq(:unlimited)
     end
 
@@ -59,7 +59,7 @@ RSpec.describe Adjective::Inventory do
     end
 
     it "will throw an error if the provided items array is longer than the max_size" do 
-      expect{Adjective::Inventory.new([@item, @item, @item, @item],{max_size: 1})}.to raise_error(ArgumentError)
+      expect{SurrogateInventory.new("Pack", 3, [@item, @item, @item, @item],{max_size: 1})}.to raise_error(ArgumentError)
     end
 
   end
@@ -152,7 +152,7 @@ RSpec.describe Adjective::Inventory do
       end
 
       it "will sort by default_sort attribute" do 
-        inventory = Adjective::Inventory.new([@item, @item, @quiver, @stick, @full_health_potion], {default_sort_method: :name})
+        inventory = SurrogateInventory.new("Pack", 4, [@item, @item, @quiver, @stick, @full_health_potion], {default_sort_method: :name})
         inventory.items.shuffle!
         ordered_inventory = inventory.sort
         expect(ordered_inventory[0].name).to eq("Healing Potion")
@@ -160,7 +160,7 @@ RSpec.describe Adjective::Inventory do
       end
 
       it "will throw an error if value for default_sort_method does not exist on all items" do 
-        expect{Adjective::Inventory.new([@item, @item, @quiver, @stick, @full_health_potion], {default_sort_method: :arbitrary})}.to raise_error(RuntimeError)
+        expect{SurrogateInventory.new("Pack", 6, [@item, @item, @quiver, @stick, @full_health_potion], {default_sort_method: :arbitrary})}.to raise_error(RuntimeError)
       end
 
     end
@@ -223,7 +223,7 @@ RSpec.describe Adjective::Inventory do
 
   context "when empty? is called" do 
     it "will return true if the inventory has no items" do 
-      @diverse_inventory.items = []
+      @diverse_inventory.dump
       expect(@diverse_inventory.empty?).to eq(true)
     end
